requirements.txt
flask
groq
python-docx
google-generativeai
requests
gunicorn
python-dotenv